pyproject.toml
[tool.poetry]
name = "apptrackr"
version = "0.1.0"
description = "Job Application Tracker"
authors = ["Devon Ho <devonho@symplesys.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-firestore = "^2.19.0"
anthropic = "^0.42.0"
streamlit = "^1.41.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
